refactor: split EXT4FileSystem.swift into 5 files; promote device contexts to library
The 1417-line EXT4FileSystem.swift had three concerns tangled
together: the FSKit pipeline methods (probe/load/unload/startCheck/
startFormat), the FFI bridge contexts wrapping FSBlockDeviceResource
and FSPathURLResource for the C-callback shape Rust expects, and the
maintenance ops + volume.info format pretty-printers.

Now split by responsibility across five files:

  DiskJockeyEXT4/EXT4FileSystem.swift   (~210 lines)
    class skeleton + MountedResource struct + statics
    (mountedResources registry, watchdog, enter/exit/scheduleWatchdog)
    + init + utility helpers (taskOption, formatCreatorOS, formatState,
    formatErrorsBehavior, formatFeatureFlags, FeatureKind) +
    MountableFileSystem conformance.

  DiskJockeyEXT4/EXT4Probe.swift        (~225 lines)
    probeResource (block-device branch), probeFileResource
    (FSPathURLResource branch), ContainerKind enum (qcow2/vhd/vhdx/
    vmdk magics), detectContainer<C: DeviceReadable> primitive.

  DiskJockeyEXT4/EXT4Load.swift         (~500 lines)
    loadResource (block-device branch), loadFileResource, unloadResource,
    didFinishLoading, buildFsCoreHandle (callbacks → optional container
    → optional partition-slice handle chain), openContainer dispatcher.

  DiskJockeyEXT4/EXT4Maintenance.swift  (~360 lines)
    The existing FSManageableResourceMaintenanceOperations extension
    (startCheck + startFormat) plus FsckProgressTracker (only used by
    startCheck, lives alongside it).

  DiskJockeyLibrary/DeviceContexts.swift (~230 lines, NEW)
    DeviceReadable protocol + BlockDeviceContext + FileDeviceContext.
    Promoted from DiskJockeyEXT4 to DiskJockeyLibrary and made public —
    these classes are FS-agnostic plumbing over FSKit + Foundation, and
    a future filesystem (e.g. NTFS file-backed mounts when those land)
    can reuse them without forking.

Method bodies move verbatim — zero behaviour change. The only
mechanical adjustments:

  - Access modifiers widen from `private` to default `internal` for the
    cross-file static helpers: detectContainer, taskOption,
    buildFsCoreHandle, openContainer, formatCreatorOS, formatState,
    formatErrorsBehavior, formatFeatureFlags. Each is `static` on
    EXT4FileSystem and reachable as `Self.X` from any extension file.

  - probeFileResource and loadFileResource stay `fileprivate` (renamed
    from `private` — Swift's `fileprivate` is the same scope at the
    type level, but `private` in an extension restricts to that
    extension block alone, which is too tight here). Their callers
    (probeResource, loadResource) live in the same file each, so the
    file-scope is sufficient.

  - DeviceReadable / BlockDeviceContext / FileDeviceContext become
    `public` so they're reachable from DiskJockeyEXT4 (which already
    imports DiskJockeyLibrary). All stored properties exposed via
    `public let`; init becomes `public`; methods become `public`.

pbxproj changes:
  - Three new file refs added to DiskJockeyEXT4 (explicit PBXGroup,
    not synchronized): EXT4Probe.swift, EXT4Load.swift,
    EXT4Maintenance.swift, with matching PBXBuildFile entries in the
    sources build phase.
  - DeviceContexts.swift slots into the synchronized DiskJockeyLibrary
    group without project edits.

No new tests for this item — pure file-split with zero new code
paths. The existing 148-test contract is the behaviour-preservation
pin:

  - DiskJockeyTests:          101/101 ✅
  - DiskJockeyLibraryTests:    33/33 XCTest + 14/14 Swift Testing ✅
  - xcodebuild build:          SUCCEEDED ✅

This completes the structural-improvement chain — items #4 (generic
FileSystemItem), #5 (FileIDCache), #6 (MountedResourceRegistry +
MountableFileSystem), #7 (MountTableParser + DiskEventHandler), #2
(RepairWatcher), and now #3 (EXT4FileSystem decomposition).
