+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The README carries an abbreviated tail of this file (last ten dated sections).
 
 ## 2026-06-02
 
+- **`EXT4FileSystem.swift` decomposed across 5 files; device contexts promoted to DiskJockeyLibrary.** The 1417-line class file had three concerns tangled together: the FSKit pipeline methods (`probeResource`, `loadResource`, `unloadResource`, `startCheck`, `startFormat`) on the orchestration side, the FFI bridge contexts (`BlockDeviceContext` for `FSBlockDeviceResource`, `FileDeviceContext` for `FSPathURLResource`) wrapping the C-callback shape `fs_ext4_blockdev_cfg_t` / `FsCoreCallbackCfg` expect, and the maintenance ops + format pretty-printers. Now split into five files by responsibility: `EXT4FileSystem.swift` (class skeleton, statics, `MountedResource`, `watchdog`, init, `taskOption` + format-display helpers, `MountableFileSystem` conformance — ~210 lines), `EXT4Probe.swift` (`probeResource` / `probeFileResource` / `ContainerKind` / `detectContainer<C>` — ~225 lines), `EXT4Load.swift` (`loadResource` / `loadFileResource` / `unloadResource` / `buildFsCoreHandle` / `openContainer` — ~430 lines), `EXT4Maintenance.swift` (the existing `FSManageableResourceMaintenanceOperations` extension + `FsckProgressTracker` — ~340 lines), plus `DiskJockeyLibrary/DeviceContexts.swift` carrying `DeviceReadable` + `BlockDeviceContext` + `FileDeviceContext` (~225 lines, made `public`) so future filesystems can reuse the same FFI bridges without forking. All method bodies move verbatim — zero behaviour change. Access modifiers widen from `private` to default `internal` for the four cross-file helpers (`detectContainer`, `taskOption`, `buildFsCoreHandle`, `openContainer`, `formatXxx`); `probeFileResource` / `loadFileResource` stay `fileprivate` since their callers stay in the same file. Three new pbxproj entries added for the EXT4 files (DiskJockeyEXT4 is an explicit PBXGroup); `DeviceContexts.swift` slots into the synchronized DiskJockeyLibrary group without project edits. No new tests for this item — pure file-split with zero new code paths; the existing 148-test contract (101 DiskJockeyTests + 33 DiskJockeyLibraryTests XCTest + 14 Swift Testing) staying green is the behaviour-preservation pin.
+
+
 - **Repair file-watcher consolidated into shared `RepairWatcher` in DiskJockeyLibrary.** `DiskJockeyEXT4/RepairXPCService.swift` (343 lines) and `DiskJockeyNTFS/RepairXPCService.swift` (204 lines) each kept a near-identical file-based-XPC scaffold — same `start()` idempotency lock, same `DispatchSource` on the requests-dir FD, same atomic-rename-as-claim, same JSON decode + filename-UUID fallback, same atomic result write. Now there's one `public final class RepairWatcher: @unchecked Sendable` in `DiskJockeyLibrary/RepairWatcher.swift` that owns all of that; each per-target `RepairXPCService` shrinks to a thin adapter that supplies only the genuine per-FS variation (`runRepair` closure body, optional watchdog enter/exit hooks, `workQueueLabel`). EXT4 keeps its ~50-line throttled-fsck `runRepair` body verbatim (with `EXT4FileSystem.watchdog.heartbeat()` and the `OperationLock` tri-state acquire/release intact); NTFS's wrapper is now ~50 lines including a 5-line stub `runRepair` that returns the truthful "not implemented yet" response. Net effect: 343 + 204 = 547 lines of duplicated scaffolding collapse to 269 lines of shared library + 130 EXT4 wrapper + 60 NTFS wrapper. NTFS's mixed log scopes (`.lifecycle` for scaffolding events, `.fsck` only inside `runRepair`) normalised to `.fsck` throughout — matches EXT4's posture and the actual meaning (repair IS fsck pipeline activity); tiny intentional log-routing tweak, not user-visible. Six new tests in `DiskJockeyLibraryTests/RepairWatcherTests.swift` drive `handleRequest(at:)` directly against temp App-Group-shaped dirs: happy-path decode + runRepair invocation + result write, malformed-JSON-with-parseable-UUID-filename writes failure result (so the host doesn't wait the full polling timeout), malformed-JSON-with-unparseable-filename writes nothing (current behaviour preserved), enter/exit hook bracket fires exactly once around handleRequest including on decode failure (critical for EXT4's watchdog counter balance), runRepair failure result is round-tripped verbatim. Full library suite 33/33 + 14 Swift Testing green; app suite 101/101 green.
 - **`AttachedDisksModel` decomposed into `MountTableParser` + `DiskEventHandler`.** The 855-line `@MainActor` model had three concerns tangled together: `/sbin/mount` subprocess + statvfs + parsing (~150 lines), structured-event → state transforms (~120 lines), and the orchestration / `@Published` store itself. The two helper concerns are now `enum MountTableParser` and `enum DiskEventHandler` in `DiskJockeyApplication/Models/` — namespace-style (case-less `enum`) so they can't be instantiated and no state can leak in. `AttachedDisksModel.refresh()` and `applyExtensionEvent` now delegate to those namespaces with one-line calls; the model's public API is unchanged. The existing `AttachedDisksModelTests` updates one call site (`AttachedDisksModel.parseMountLine` → `MountTableParser.parseMountLine`) — that the test still passes after the rename is the proof of behaviour preservation. New `DiskJockeyTests/MountTableParserTests.swift` (11 cases) covers parseMountLine for RW / RO / legacy-RO / fstype-filter / malformed-line / NTFS, bsdName extraction (both dev-prefixed and bare), and isWholeDiskBSD on bare disks / slices / malformed input including the `disk` edge case the original regex rejects. New `DiskJockeyTests/DiskEventHandlerTests.swift` (17 cases) covers fsTypeFromEventKind for known and unknown prefixes, fsckStatus dispatch across every kind (volume.clean / dirty / fsck.start / fsck.progress / fsck.done / fsck.failed / unknown), applyVolumeInfo (fs adoption, volume_name placeholder rule, volume_name clobber-protection, stableIdentity from UUID + from ntfs serial), materializeTotalSize (ext4 multiply / ntfs passthrough / unknown / overflow), and applyEventInPlace dispatch (volume.info → volume handler / fsck.progress → fsckStatus / unknown kind no-op). Full app + library suites still green (139 tests across both).
 - **`MountableFileSystem` protocol + `MountedResourceRegistry<R>` generic.** EXT4FileSystem and NTFSFileSystem each declared `static let mountedResources = OSAllocatedUnfairLock<[ObjectIdentifier: MountedResource]>(initialState: [:])` and reimplemented the same four ops against it: register on load, remove on unload, resolve-the-sole-entry for `startCheck`/`startFormat`, and find-by-bsdName for the RepairXPCService. Same shape, two byte-near-identical implementations. Now there's one `public final class MountedResourceRegistry<Resource: MountedResource>: @unchecked Sendable` in `DiskJockeyLibrary/MountableFileSystem.swift`, exposing `register(_:_:)` / `remove(_:)` / `resolveSingle()` / `first(where:)` / `count` — the four operations the callers actually do, each one line at the call site. Plus `protocol MountedResource` (codifies the shared `bsdName + opLock` fields the per-FS structs already had) and `protocol MountableFileSystem` (associates each FS class with its registry handle). Twelve raw `withLock { map in … }` blocks collapse to one-line API calls; both `EXT4FileSystem.MountedResource` / `NTFSFileSystem.MountedResource` adopt the library protocol with zero field changes; both classes carry an `extension … : MountableFileSystem {}` conformance. Eight new tests in `DiskJockeyLibraryTests/MountableFileSystemTests.swift` cover register / replace / remove / idempotent-remove / `resolveSingle` ambiguity refusal / `first(where:)` predicate / concurrent 100-key register-and-remove / a stand-in `MountableFileSystem` conformance that proves the associated-type machinery compiles independent of the FSKit-extension targets. Full library suite 41/41 green.
